型整理
