+++ lib/search.ts
@@ -1,15 +1,11 @@
-import { RecipeType } from '~/types/recipe'
+import { Response } from '~/types/recipe'
 
-export type Response = {
-  recipe: RecipeType
-}
-
-export async function getRecipe(id: number): Promise<Response> {
+export async function search(word: string): Promise<Response> {
   const headers = {
     'X-Api-Key': process.env.API_KEY ? process.env.API_KEY : '',
   }
   const response = await fetch(
-    `https://internship-recipe-api.ckpd.co/recipes/${id}`,
+    `https://internship-recipe-api.ckpd.co/search?keyword=${word}`,
     {
       method: 'GET',
       headers: headers,
